+++ project2group6/src/index.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
 import reportWebVitals from './reportWebVitals';
-import List from './List';
+import List from './pages/List';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
